requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.2
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
drf-nested-routers==0.93.4
idna==3.4
kavenegar==1.1.2
Pillow==10.0.0
psycopg2==2.9.6
python-decouple==3.8
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.4
